Adds better support for custom data indicators
When resolving a default consolidator we'll now always use identity IF the resolution was not specified or if it matches the subscription's resolution. The idea here being if they're the same then we can safely assume that we don't need to aggregate/transform the data in any way and just pipe the .Value to the indicator for every piece of data we get.

This allows the 1D short code indicator functions to work (EMA, SMA, MACD, RSI, ect..)