Update SetStart/EndDate functions to be full days
Using the SetEndDate function would end up with the last day of data not being processed due to a period start/finish filter in the SubscriptionDataReader.MoveNext() function. Updating the SetEndDate to make the DateTime value to last moment in the day solves the issue. It also ends up fixing a lingering bug in the ConsoleResultHandler where we were computing the wrong number of days to process.