Adds Composer<T> for accessing MEF'd instances
As more of the QC application moves towards using MEF for resolving certain types and plugins at run time, it seems necessary to provide a single type that can do the composition and also allow querying of the composed instances. Further, it maintains the composed instances for the rest of the application to access.