OBSOLETES:: BaseData.Reader and BaseData.GetSource
The existing method signatures have been made obsolete in favor of something slightly simpler and more deterministic.
The new methods accept a bool isLiveMode flag instead of the DataFeedEndpoint.
The GetSource method returns a new type: SubscriptionDataSource, which combines the source string and the SubscriptionTransportMedium.

Existing sub-types of BaseData have been updated with the new signatures.  IBaseData has not been touch as it is mostly unused by LEAN infrastructure.