Adds SymbolJsonConverter
For some still unknown reason, when trying to deserialize a
List<Ticks> into a List<BaseData> with multiple symbols, all
instances would share the same symbol reference (to the last
symbol in the list). Adding a dedicated json converter to
handle the deserialization handled this.

SymbolTests.HandlesListTicksWithDifferentSymbols has the
demonstrating case of the bug.