Rewrites the LiveTradingDataFeed to be enumerator centric
This refactor was an effort to bring into line the various concepts between the
FileSystemDataFeed and the LiveTradingDataFeed. The former works using enumerators
and a time sync loop that uses the concept of a frontier to decide when to 'pull-off'
data that's at or before the frontier. This makes the feed uninterested in how the
data (via enumerators) is provided, and only concerns itself with things at the
subscription level, that is, time syncing and universe selection invocation.

These concepts were brought over to the LiveTradingDataFeed with some modification.
This change heavily uses object composition of enumerators to handle the various
concerns that were previously within the enumerator loops in the LiveTradingDataFeed.

Some enumerator types that help accomplish these concerns:

	FastForwardEnumerator 		- fast forwards an enumerator that contains old data
	RateLimitEnumerator			- prevents an enumerator from being invoked too frequently
	TradeBarBuilderEnumerator 	- Builds trade bars from tick data
	FrontierAwareEnumerator		- Emits the underlying when the frontier is on or after Current
	EnqueableEnumerator			- Acts as a liason between a push/pull system using a queue