Prevent brokerages from sending 0 holding positions
IBrokerage.GetAccountHoldings() is intended to help initialize the algorithm state.
We use the results from this to add extra securities to the algorithm. In some cases
the brokerages (specifically IB) will send back account positions with zero holdings,
this still causes us to add the security, and in the event it's an option holding,
it will prematurely kill the algorithm as we don't support options yet.