Fixes test proj dependency issue
QuantConnect.Tests seems to have a transitive dependency on IKVM.OpenJdk.Jdbc
The Composer was unable to load dlls without it. The compiler was not copying
the file over since it doesn't appear to be required, but possibily
something in the Composer code is trying to load an assembly which does require
it.

The short fix here is to make a static, compile time reference to the assembly
from within the Fxcm brokerage implementation. This will ensure that the compiler
properly copies this dlls into the various output directories.