Oanda DataQueueHandler streams no data with more than one Subscribe request
Oanda only supports a maximum of two streaming rates sessions per token so we are reusing the same session for all subscriptions. In the case of multiple symbols to subscribe, on subsequent requests we have to close the previous session and start a new one, so requests must be synchronous.