Fixes IB partial fill bug
It turns out that we couldn't trust the update.Status field.
We were looking at this to get the correct status for the fill,
but even when update.Filled!=0, the status may be Presubmitted,
or just Submitted. In testing I was unable to detect an update.Status
==PartiallyFilled, so we determine PartiallyFilled/Filled status
based on the internal state of the IB brokerage implementation