Update Universe selection to always add selected members
This behavior was always dubious at best. The goal was to not add a security
to a universe as a 'member' if we couldn't get data for it. These two are
separate issues though. First the security becomes a member, and then we
check for its data. Also, it could (and should) be possible to add a security
to a universe and not add data for (since maybe another universe already
added data for that security)