Fixes ApplyUniverseSelection for adding existing member
As long as the subscription is not an universe, all symbols selected to be added were considered a new addition, even when those symbols were already part of that universe.
Now we only update our security changes if we actually added a new member to the universe after we succeed to subscribe it.