QuantConnect.Views is now a console application
In order to decouple the build profiles of QuantConnect.Views and the rest of the application, QuantConnevt.Views will now produce an executable that can be called from the rest of the project.  All references to QuantConnect.Views from the Laucher project are now removed.  This is all done so that The rest of the project can compile into x64.  QuantConnect.Views relies on the Gecko web browser which only comes in 32 bit flavor.