LeanWinForm.cs now implements IMessageHandler
The WinForm application now implements the interface that allows the DesktopClient to pass messages to it. This replaces the delegate methods that were being used before.

The Main method in the Views project now starts the NetMQ pull socket (server) in another thread.  This is so the server does no block the rest of the WinForms application.  The server is then shut down once the application is finished running.