Minor changes to URI construction and API Initialization.
In order to better handle future scenerios, a new URI is constructed from the data download link returned from the API.  The URI host, scheme and query+path is used to reconstruct the link to download the data.

In addition, the API Initialize() method now takes the dataFolder as a parameter and saves it as a private variable  so as to not read the data folder from the config every time.