Add DisposeSafely extension methods
Add DisposableExtensions to hold extension methods for IDisposable.
Add two overloads of DisposeSafely. This will allow you to call
myDisposable.DisposeSafely() and it will wrap the call in a
try/catch and log the error. The other overload allows you to
handler the error using an Action<Exception> errorHandler parameter