Swallow bug in QuoteBar.Reader() if an error is thrown during parsing
Following the pattern in TradeBar.Reader(), the QuoteBar.Reader() method will now swallow a parsing error, log the error message and line number and then proced to return a default instance of Quotebar.  The current implementation of QuoteBar.Reader() will stop the executation of the program if a parsing error is found.  Given the amount of data Lean may parse during the execuation of an algorithm, it is better to be resiliant to small errors in the data than to stop the execuation of the program because of a malformated line of csv.