Fixes Cash.EnsureCurrencyDataFeed to enable Oanda and QuoteBar
Cash.EnsureCurrencyDataFeed was looking subscribing to currencies with trade bar information while they come as quote bars.
Since it was using DefaultMarkets to select the market the symbol belongs to, it was subcribing to currencies from the FXCM market despite the fact the algorithm could be using Oanda data. Wih this fix, when the algorithm is trading CFD or Forex it will subscribe to a quote currency from the same market.