Fixes ib pacing flag
As written, once this flag is set to true, then it will always
be true for every subsequent iteration of the loop. IOW, once
you get pacing violation then it will always interpret timeouts
as a pacing violation via line 2079. The fix here moves the
declaration of the flag to inside the loop, so that on each
iteration we're guaranteed that it's reset to the false state.