Refactor GroupByAdjacent into overload
This new overload accepts just the enumerator instead of the full enumerable.
This enables easy usages with the enumerator stack, specifically the
SynchronizingEnumerator.

In addition, the method was updated to return an IList<T> to represent the inner
grouped collection. The current implementation of this method is such that it
emits only full groupings, so this return type change codifies that implementation
detail into the type signature, allowing consumers to depend on it.