WIP: Initial implementation of syncing IB GetHistory
I threw this together this evening as a proof of concept of what I'm going for
here. Each request results in a BaseDataCollection time stamped in the slice time
zone. The separate collections are then synchronized by the SynchronizingEnumerator
and formed into Slice instances.

I've modified the time zone handling of the raw data. It appears as though we're
modifying the time zone of the data to match the requested sliceTimeZone, but
really the data should get fed to the algorithm in the configured exchange time
zone.