Added LeanData.TryParsePath() and associated tests
This method takes a path and will return a Security and DateTime.  This is useful in the ApiDataProvider that uses the api to retrieve files that are not present on disc.  The IDataProvider only ha a path to the file.  Using this method, the ApiDataProvider can extract the useful information from the path to communicate with the api.