Fix quantity bug in option exercise/assignment
When assigning shares of the underlying equity for an option at expiration, the equity quantity was incorrectly set equal to the option quantity. Now this quantity is multiplied by option.ContractUnitOfTrade (usually 100 for US options).

This bug was hidden by the option assignment tests in SecurityPortfolioManagerTests, which have also been corrected.
The regression stats for OptionExerciseAssignRegressionAlgorithm have also been updated.

In some cases this bug was also causing division by zero exceptions in OnOrderEvent, which seem to be fixed with this PR.