Add null check in BRH.ProcessSynchronousEvents()
There are cases were creating the instance of the algorithm or setting up the algorithm will throw an error. In those cases, the backtesting result handler - who has already been initiaized - will attempt to call ProcessSynchronousEvents() when it's instance of IAlgorithm is still null. This will then throw an error when BRH.ProcessSynchronousEvents() attempts to access the properties on IAlgorithm