Truncated the EnsureCurrencyDataFeed
this method was trying to load currency conversion rates for every currency any of my IB assets are denominated in. I had some USD in my IB account, despite my IB account being based in AUD so to avoid an attempt to load a conversion rate for those usd assets i had to remove a large part of the EnsureCurrencyDataFeed method.