Safely dispose of parallel workers
Dispose on the parallel controller is only called in backtest at the
very end of the backtest. Popping the stack due to an error in disposing
a resource that is about to be cleanedup by program exit seems like a
poor clean up strategy.

By using DisposeSafely, we'll completely ignore ObjectDisposedExceptions
and we'll simply log any other exceptions but allow the program flow to
continue. This will allow other components to be diposed downstream in
the event an error s raised.