Make subscription enumeration order deterministic
In the SubscriptionCollection class, subscriptions are stored in nested dictionaries, keyed by Symbol and by SubscriptionDataConfig.
According to MSDN when enumerating: The order in which the items are returned is undefined.
https://msdn.microsoft.com/en-us/library/xfhwa508.aspx

The enumeration ordering for different symbols is not an issue, but the order of configs for the same symbol can cause non-deterministic backtest results (only security types with multiple data types per subscription are affected such as Crypto, Futures, Options). This random behavior was observed with the FractionalQuantityRegressionAlgorithm, so the regression statistics have also been updated.

In this PR we force a fixed priority for data types, which will guarantee that different data points for the same symbol at the same time step will always be emitted in the same order (TradeBar before QuoteBar).