Add IBrokerageModel.GetMarginModel and invoke from BrokerageModelSecurityInitializer
It seems all the other models are being provided via the brokerage model but all the margin models
are assigned via the security constructors. This allows us to set margin models using the same
pattern that we used for overriding default margin models. This becomes particularly important
when considering how different exchanges/brokerages manage cash and margin in the account.