Fix out-of-memory and Too Many Open Files errors when running Lean under Mono in Linux.
Because the C# garbage collector is implemented differently in Mono under Linux than in Windows, we have found that running on very fast systems as well as running with very large data files can lead to either a Too Many Open Files error or seg-faults as the system runs out of memory.

The solution to this is two-fold:

1)  There's no reason that it should be necessary to introduce the non-determinism of periodically running CleanCache.  Instead, let's just run CleanCache when we actually need it.  This will likely be more efficient than the previous method.
2)  Limit the number of simultaneously open, cached files allowed.

It is noteworthy that even removing the MemoryStream and letting ZipInputStream handle all I/O does not fix this issue.