Move security type ordering to SubscriptionCollection
This will change the order in which we enumerate subscriptions so we don't need
to re-sort inside of TimeSlice.Create.

The way we currently build time slices requires that we receive the underlying data
before any derivative data. This change ensures that we enumerate the underlying equity
data before we process any derivative data, such as options.