Check market hours in ImmediateFillModel
Previously we were seemingly reliant on not receiving data to prevent fills
from ocurring outside of market hours. This breaks down when using multiple
security type with different market hours and caused fills to be generated
outside of market hours using the closing price from the previous day.

This change now explicitly checks to confirm that the market is open before
creating the order fill. Also, this change allows limit/stop limit orders to
fill in pre/post market but forces market/stop market orders to only fill
within normal market hours. This behavior can be brokerage specific (some
allow pre/post market trading, others do not) and as such, a brokerage model
can define it's own fill model to use to alter the default behavior
described here.

The changes to the LimitFillRegressionAlgorithm are due removing out of
market fills.

The changes to SecurityTransactionModelTests are to properly set the start
and end times of the trade bar data used when setting the market price. This
allows us to properly check market hours.