Add short selling pre-order check
Currently, when using a cash model and calling SetHoldings with a negative percentage (to go short), CalculateOrderQuantity is returning a quantity of zero as expected, so no order is submitted and no error or warning message is shown.

In this PR, the IsShortSellingAllowed method has been added to the IBrokerageModel interface, all order methods call it and an error message is shown if the short position is not allowed.