Add support or reference result types in TypeChangeJsonConverter
Existing implementation was written under the assumption that the result type
was a value type and was being read using JToken.Value<TResul>() which cases
exceptions if TResult is a non-string class type. The fix was to check the type
and use JToken.ToObject<TResult>() when the result type is not a value type.