Add Contracts function to OptionFilterUniverse
OptionFilterUniverse was lacking a means of explicitly setting the set of
contract symbols to be selected, which means algorithms had to use the
various helper methods but were unable to define their own selection function
for the contract universe. This change adds two methods, one to epxlicitly
set the contracts to be selected and another that mirrors the way universe
selection works, by providing a function that accepts the full set of contract
symbols and returns the filtered set of contract symbols. This enables LINQ
filtering of the contract universe.