First commit for Kraken API
First commit for Kraken as it is. Have to re-check everything (including coding style) and fix lots of things. Tests not working / done, and actually API code had 0 testing.

Original author of C# Kraken Rest API code is trenki2
https://github.com/trenki2
It is being adapted so that it will work with Lean engine.