Fix IEXDataQueueHandler data feed
When selecting IEXDataQueueHandler as a data source for live trading,
it would not connect. The following fixed the issue:

- Updated Socket IO NuGet package (and its dependencies) to latest;

- In GzipStreamProvider, removed obsolete "using" statement refering
to deprecated Socket IO API (using statement was not used anyways);

- Increased the reconnection delay value to its recommended
value (1000ms), this was necessary for the feed to connect;

- Minor change in IEXDataQueueHandler to parse Json message.
"(JObject)message" would compile, but threw a runtime error, as this
formulation could not implicitly convert string to JObject. Used
the cleaner "JObject.Parse()" method instead.