Implement IEX history provider
Implemented IHistoryProvider interface in class IEXDataQueueHandler,
to make it usable for history requests as well as real-time data feed.

Note that due to IEX implementation, only daily resolution history
requests are supported, up to a maximum of 5 years look back.

Implemented tests related to history requests.

To demonstrate how to use it, modified config.json to propose a
working environment using IEX as a history provider, and as a live
data feed.