Updates DockerfileJupyter
Updates the link to fetch a new version of iCSharp with Jupyter message specification v5.

The new docker image places Python.Runtime.dll from the nuget package in pythonnet folder as this dll is used in C# notebooks.
Python notebook use Python.Runtime.dll that is compiled in docker when pythonnet is installed via pip.