Previous version entered Nifty OnData method prior to initializing CorrelationPair in DollarRupee method
Made this adjustment to ensure that CorrelationPair is properly instantiated -- previous method didn't initialize it in DollarRupee OnData method and so failed in evaluation later. Also, the Python version of this algo. used Portfolio.MarginRemaining to calculate quantity and so I changed this one to ensure they return the same results.