Update Future.PriceMultipliers table
The previous table was outdated, many symbols are not being traded.

The new table was generated using the the data available in [CME Group page](https://www.cmegroup.com/trading/products/index.html)

The criteria to select the securities was filtering the securities with zero volume or zero open interest.