Remove start date from coarse generator
This commit removes start date in the coarse generator. This option was used in the previous version, in order to process only the newest dates.

The reason is the fact that coarse files now contains factors to estimate adjusted prices. In turn, factors are updated backward with a new corporate event (split, dividends, etc); the new coarse generator should *always* process the full symbol historical data, for all symbols.