LEAN-3631: Resolve non-fill-forward data in TimeSlice.Create
Using LINQ to filter millions of these lists takes forever and ends up allocating
hella lots of objects to the heap. By doing it this way we make the list once as
part of an existing loop that is already CPU intensive, so it blends right in with
the other stuff vs having the hiccup in the middle of the algo manager time  loop.