Correct implementation of the first row issue
in the previous solution, I basically tried to pass the first row as a split of 1.

That's is not a correct solution, the correct one is treat the first row as a different  _kind_ of row. In fact, the the main usage of the first row is a date reference for the factors defined in the second row