Fixes incorrect beta calculation
* Reworks Sampling portion of AlgorithmManager loop
* Fixes issue with miscalculation of various portfolio statistics
* Fixes data alignment issue in statistics calculation

We now truncate the first few values of the "Daily Performance"
and benchmark performance series in order to align the series
correctly and attain accurate statistics at the cost of reducing
our sample size by 1 or 2 in the StatisticsBuilder class.

The shape of the data has also changed due to these changes.
Changes to the shape are:
  * The zero at the end of the "Daily Performance" series is gone
  * "Daily Performance" series now begins with zero.

We also now sample at the end of the AlgorithmManager loop to ensure
that the sampling events are fired even if we have no remaining points
left. This means that sampling after the Stream(...) loop is no longer
needed (if the algorithm successfully ran to the end), though we do
still force sample at the end just in case.

And finally, we've removed the Sample...(...) methods from the
AlgorithmManager and are now sampling inside the
`results.ProcessSynchronousEvents()` call.