Fixes Stop Market Fill of Equity Fill Model
Only use trade data (Tick with trade type or TradeBar) to get the high and low prices, since stop market orders are triggered with high and low price. The fill price is the worse case scenario between the stop price and the bid/ask, unless there is a gap and the open price is used instead.
https://www1.interactivebrokers.com/en/index.php?f=609

Fix unit tests to show that the new implementation only fills with high and low values of trade data.

Change StopLossOnOrderEventRegressionAlgorithm regression tests to reflect the bug fix. The change results from the price gap handling as the both TradeBar and QuoteBar high and low triggers the stop loss.