Fixes Stop Market Fill of Equity Fill Model
Use the latest `TradeBar` `High` or `Low` (instead of `QuoteBar`) to trigger the stop market fill and fill with the worst-case scenario between the stop price and the bid or ask. Unless the bar opens above/below the stop price. In this case, we fill at the open. For tick resolution, use the maximum or minimum of the last trades.