Updates Unit Tests: Fills with Quote
If the trigger price and the limit price are touched with the same trade bar (or batch of tick trades), we cannot use the high and low because we cannot ensure that the trigger price was touched before the limit price. In this case, we need to consider the closing (quote) prices to trigger the limit order.