Handles Trigger and Fill On the Same Bar
There are four types of Limit-if-touched (LIT) orders:
1. Marketable limit price
  a. Buy  LIT order: set the trigger price below the current market price and set the limit price above the trigger price.
  b. Sell LIT order: set the trigger price above the current market price and set the limit price below the trigger price.
2. Unmarketable limit price
  a. Buy  LIT order: set the trigger price below the current market price and set the limit price below the trigger price.
  b. Sell LIT order: set the trigger price above the current market price and set the limit price above the trigger price.

Therefore, the trigger price can be lower or higher than the limit price, and it's uncertain that the trigger price is attained or penetrated before the limit price with bar information. Consequently, when the trigger price is touched for the first time, the closing price (quote) will test the limit price.

This fill model handles the LIT fill as follows:
1. If the trigger price is attained or penetrated, the order flag TriggerTouched is set to true.
2. On the first touch, the limit order is filled if the closing price penetrates the limit price.
3. When new trade information arrives, the LIT order is handled as a limit order: the limit order is filled if the high/low penetrated the limit price.