Adds Unit Tests to Assert Expected Behavior
The stop limit method should not fill using QuoteBar or Tick with TickType.Quote type unless the trigger price and the limit price are potentially attained or penetrated in the same bar. In this case, we need to use the closing (quote) price to test the limit fill since it's certain that the closing price doesn't come before the trigger price.

Adds missing test for tick data (TickType.Trade).

Favorable gap test (see #963).