Increase miniconda lib priority in load paths
The OpenSSL library requested by python311 in this configuration (3.5.5)
conflicts with what is available in /usr/lib (3.0.2), making it
impossible to use any SSL functionality in python.
Update the definition of LD_LIBRARY_PATH to include /opt/miniconda3/lib/
and make sure it's top of the list so that libraries are searched there
first.
Also, drop the reference to the undefined LD_LIBRARY_PATH.

Signed-off-by: Alessio Balsini <alessio.balsini@gmail.com>